Cleanup compiler warnings

+++ CloudSeed/AllpassDiffuser.h
@@ -23,7 +23,7 @@ namespace CloudSeed
 		vector<float> seedValues;
 		int seed;
 		float crossSeed;
-		
+
 	public:
 		int Stages;
 
@@ -95,7 +95,7 @@ namespace CloudSeed
 			return filters[Stages - 1]->GetOutput();
 		}
 
-		
+
 		void SetDelay(int delaySamples)
 		{
 			delay = delaySamples;
@@ -110,7 +110,7 @@ namespace CloudSeed
 
 		void SetModAmount(float amount)
 		{
-			for (int i = 0; i < filters.size(); i++)
+			for (size_t i = 0; i < filters.size(); i++)
 			{
 				filters[i]->ModAmount = amount * (0.85 + 0.3 * seedValues[MaxStageCount + i]);
 			}
